+++ docs/keyboard_shortcuts.txt
