+++ scripts/main.spec
